+++ epstein_files/util/helpers/rich_helpers.py
@@ -1,13 +1,20 @@
 """
 Small methods for building rich text constructs.
 """
+import os
+
 from rich.padding import Padding
 from rich.text import Text
+
+from epstein_files.util.helpers.string_helper import starred_header
 
 left_indent_padding = lambda num_spaces: (0, 0, 0, num_spaces)
 left_indent = lambda obj, num_spaces: Padding(obj, left_indent_padding(num_spaces))
 no_bold = lambda style: style.replace('bold', '').strip()
+suppress_output_console_kwargs = lambda: {'file': open(os.devnull, "wt")}
 vertically_pad = lambda obj, amount = 1: Padding(obj, (amount, 0, amount, 0))
+
+WARNING_STYLE = 'bold black on white'
 
 
 def enclose(txt: str | Text, encloser: str = '', encloser_style: str = 'wheat4') -> Text:
@@ -25,3 +32,8 @@ def join_non_empty(*txts, sep: str | Text = ' ') -> Text:
     """Join two strings but only if they are not empty."""
     sep = sep if isinstance(sep, Text) else Text(sep)
     return sep.join([t for t in txts if len(t) > 0])
+
+
+def starred_header_txt(msg: str, num_stars: int = 7, num_spaces: int = 2, style: str = WARNING_STYLE) -> Text:
+    """String like '  *** Title Msg ***  '."""
+    return Text(starred_header(msg, num_stars, num_spaces), style)