+++ tests/output/site/test_sites.py
@@ -4,5 +4,5 @@ from epstein_files.output.site.sites import *
 
 
 def test_build_path():
-    assert SiteType.build_path(SiteType.CHRONOLOGICAL) == HTML_DIR.joinpath('index.html')
-    assert SiteType.build_path(SiteType.DEVICE_SIGNATURES) == HTML_DIR.joinpath('device_signatures.html')
+    assert SiteType.html_output_path(SiteType.CHRONOLOGICAL) == HTML_DIR.joinpath('index.html')
+    assert SiteType.html_output_path(SiteType.DEVICE_SIGNATURES) == HTML_DIR.joinpath('device_signatures.html')