+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* use the `DocPrinter` to build the title page so it correctly centers in the browser
 * `DocCfg.highlight_quote` now works with `Email`s too
 * `Document._config` method falls back to a dummy / empty `DocCfg` / `EmailCfg` / etc.
 * `DocCfg` new property `num_preview_chars`, standardize preview length computation